Break out of matcher/any.hpp
