+++ include/trompeloeil/matcher.hpp
@@ -205,44 +205,6 @@ namespace trompeloeil
       matcher_kind_t<MatchType, Predicate, detail::decay_t<T>...>,
       detail::decay_t<T>...>;
 
-  namespace lambdas {
-
-    struct any_predicate
-    {
-      template <typename T>
-      bool
-      operator()(
-        T&&)
-      const
-      {
-        return true;
-      }
-    };
-
-
-    // Define `struct` with `operator()` to replace generic lambdas.
-
-    struct any_printer
-    {
-      explicit
-      any_printer(
-        char const* type_name_)
-        : type_name(type_name_)
-      {}
-
-      void
-      operator()(
-        std::ostream& os)
-      const
-      {
-        os << " matching ANY(" << type_name << ")";
-      }
-
-    private:
-      char const* type_name;
-    };
-
-  }
 
   template <typename MatchType, typename Predicate, typename Printer, typename ... T>
   inline
@@ -253,39 +215,7 @@ namespace trompeloeil
   }
 
 
-  template <
-    typename T,
-    typename R = make_matcher_return<T, lambdas::any_predicate, lambdas::any_printer>>
-  inline
-  auto
-  any_matcher_impl(char const* type_name, std::false_type)
-  TROMPELOEIL_TRAILING_RETURN_TYPE(R)
-  {
-    return make_matcher<T>(lambdas::any_predicate(), lambdas::any_printer(type_name));
-  }
-
-  template <typename T>
-  wildcard
-  any_matcher_impl(char const*, std::true_type);
-
-  template <typename T>
-  inline
-  auto
-  any_matcher(char const* name)
-  TROMPELOEIL_TRAILING_RETURN_TYPE(decltype(any_matcher_impl<T>(name, std::is_array<T>{})))
-  {
-    static_assert(!std::is_array<T>::value,
-                  "array parameter type decays to pointer type for ANY()"
-                  " matcher. Please rephrase as pointer instead");
-    return any_matcher_impl<T>(name, std::is_array<T>{});
-  }
 
 }
-
-#define TROMPELOEIL_ANY(type) ::trompeloeil::any_matcher<type>(#type)
-
-#ifndef TROMPELOEIL_LONG_MACROS
-#define ANY                       TROMPELOEIL_ANY
-#endif
 
 #endif //TROMPELOEIL_MATCHER_HPP
